MeIN-docker-demo/django/requirements.txt
Django==3.1.4
Pillow
jsonify
requests
cryptography
gunicorn==20.0.4
djangorestframework
djangorestframework-jwt
django-filter
django-rest-auth
django-allauth
asgiref
pytz
sqlparse
drf_yasg
django-cors-headers
firebase-admin
django-background-tasks